+++ lib/monitor/chart.g.dart
@@ -1,0 +1,24 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'chart.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$pointsHash() => r'388c70dd677518225418cb658e7dd8fe223b05ab';
+
+/// See also [_Points].
+@ProviderFor(_Points)
+final _pointsProvider =
+    AutoDisposeNotifierProvider<_Points, List<FlSpot>>.internal(
+  _Points.new,
+  name: r'_pointsProvider',
+  debugGetCreateSourceHash:
+      const bool.fromEnvironment('dart.vm.product') ? null : _$pointsHash,
+  dependencies: null,
+  allTransitiveDependencies: null,
+);
+
+typedef _$Points = AutoDisposeNotifier<List<FlSpot>>;
+// ignore_for_file: unnecessary_raw_strings, subtype_of_sealed_class, invalid_use_of_internal_member, do_not_use_environment, prefer_const_constructors, public_member_api_docs, avoid_private_typedef_functions